新版脚本用旧版解释器会提醒更新到最新版本

+++ firefox-91.sh
@@ -1,4 +1,11 @@
 #!/usr/bin/env deepin-wine-runner-auto-install-bash
+# 用于判断是否为 bash 解释器
+# 旧版解释器直接提示不兼容
+if [[ 1 == 2 ]]; then
+    error 错误 此脚本只支持\ Wine\ 运行器\ 2.3.0\ 及以上版本，请更新至最新版本
+    exit
+fi
+# 操作
 echo Power By $MAKER
 echo $COPYRIGHT
 rm -rfv /tmp/firefox.exe
